otlp2parquet_core::otlp::common::format

Trait OtlpSignalRequest

pub trait OtlpSignalRequest:
    Message
    + Default
    + DeserializeOwned {
    const JSONL_EMPTY_ERROR: &'static str;

    // Required methods
    fn merge(&mut self, other: Self);
    fn is_empty(&self) -> bool;
}

Trait implemented by OTLP signal request types that can be parsed from multiple formats.

Required Associated Constants

const JSONL_EMPTY_ERROR: &'static str

Error message returned when JSONL input does not contain any usable records.

Required Methods

fn merge(&mut self, other: Self)

Merge another request into self.

fn is_empty(&self) -> bool

Returns true if the request does not contain any signal data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.